# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Welcome to family budget API.",
        "title": "Family Budget API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Isfandiyor Zabirov",
            "url": "http://www.swagger.io/support",
            "email": "isfandiyor.zabirov.sh@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "10.192.9.236:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/goals": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Получает список целей по фильтрам",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Цели"
                ],
                "summary": "Получение списка целей",
                "operationId": "get-goals-list",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Поиск по названию и описанию",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Статус цели",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата выполнения 'от' (YYYY-MM-DD)",
                        "name": "due_date_from",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Дата выполнения 'до' (YYYY-MM-DD)",
                        "name": "due_date_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы",
                        "name": "current_page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Количество на странице",
                        "name": "page_limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Обновляет данные существующей цели по ее ID из тела запроса",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Цели"
                ],
                "summary": "Обновление цели",
                "operationId": "update-goal",
                "parameters": [
                    {
                        "description": "Данные для обновления цели (включая ID)",
                        "name": "{object}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/goals.Goals"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "404": {
                        "description": "Цель не найдена",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Создает новую финансовую цель для семьи",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Цели"
                ],
                "summary": "Создание новой цели",
                "operationId": "create-goal",
                "parameters": [
                    {
                        "description": "Данные для создания цели",
                        "name": "{object}",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/goals.Goals"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "Неверные входные данные",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            }
        },
        "/goals/{id}": {
            "get": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Получает детальную информацию о цели по ее ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Цели"
                ],
                "summary": "Получение информации о цели",
                "operationId": "get-goal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Цели",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "404": {
                        "description": "Цель не найдена",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWT": []
                    }
                ],
                "description": "Удаляет цель по ее ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Цели"
                ],
                "summary": "Удаление цели",
                "operationId": "delete-goal",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Цели",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "400": {
                        "description": "Неверный формат ID",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "403": {
                        "description": "Доступ запрещен",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "404": {
                        "description": "Цель не найдена",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/response.ResponseModel"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "goals.Goals": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "due_date": {
                    "type": "string"
                },
                "family_id": {
                    "description": "к какой семье принадлежит",
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "remaining_budget": {
                    "description": "сколько бюджета осталось",
                    "type": "number"
                },
                "status": {
                    "type": "string"
                },
                "total_budget": {
                    "description": "сколько бюджета нужен для выполнение цели",
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "response.ResponseModel": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}